Выполнить оптимизацию запросов в сервисе категорий и подборок: ввести индексы, решить проблемы N+1 запросов. Выполнить разделение интерфейсов в event-service согласно принципам SOLID.
