+++ core/event-server/src/main/java/ru/eventlink/category/controller/CategoryPublicController.java
@@ -7,7 +7,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import ru.eventlink.category.mapper.CategoryMapper;
-import ru.eventlink.category.service.CategoryService;
+import ru.eventlink.category.service.CategoryPublicService;
 import ru.eventlink.dto.category.CategoryDto;
 
 import java.util.List;
@@ -18,7 +18,7 @@ import java.util.List;
 @Slf4j
 @Validated
 public class CategoryPublicController {
-    private final CategoryService categoryService;
+    private final CategoryPublicService categoryPublicService;
     private final CategoryMapper categoryMapper;
 
     @GetMapping
@@ -27,13 +27,13 @@ public class CategoryPublicController {
                                               @RequestParam(defaultValue = "10")
                                               @Positive Integer size) {
         log.info("getAllCategories: page={}, size={}", page, size);
-        return categoryMapper.listCategoryToListCategoryDto(categoryService.getAllCategories(page, size));
+        return categoryMapper.listCategoryToListCategoryDto(categoryPublicService.getAllCategories(page, size));
     }
 
     @GetMapping("/{catId}")
     public CategoryDto getCategory(@PathVariable Long catId) {
         log.info("getCategory: catId={}", catId);
-        return categoryMapper.categoryToCategoryDto(categoryService.getCategory(catId));
+        return categoryMapper.categoryToCategoryDto(categoryPublicService.getCategory(catId));
     }
 
 }
