+++ core/event-server/src/main/java/ru/eventlink/category/repository/CategoryRepository.java
@@ -1,5 +1,6 @@
 package ru.eventlink.category.repository;
 
+import jakarta.validation.constraints.NotBlank;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.eventlink.category.model.Category;
 
@@ -7,4 +8,6 @@ import java.util.Optional;
 
 public interface CategoryRepository extends JpaRepository<Category, Long> {
     Optional<Category> findCategoriesByNameContainingIgnoreCase(String name);
+
+    boolean existsByNameContainingIgnoreCase(@NotBlank String name);
 }
