+++ core/event-server/src/main/java/ru/eventlink/category/service/CategoryPublicService.java
@@ -1,0 +1,11 @@
+package ru.eventlink.category.service;
+
+import ru.eventlink.category.model.Category;
+
+import java.util.List;
+
+public interface CategoryPublicService {
+    List<Category> getAllCategories(int page, int size);
+
+    Category getCategory(long catId);
+}
