+++ core/event-server/src/main/java/ru/eventlink/compilation/service/CompilationAdminService.java
@@ -4,16 +4,10 @@ import ru.eventlink.dto.compilation.CompilationDto;
 import ru.eventlink.dto.compilation.NewCompilationDto;
 import ru.eventlink.dto.compilation.UpdateCompilationRequest;
 
-import java.util.List;
-
-public interface CompilationService {
+public interface CompilationAdminService {
     CompilationDto addCompilation(NewCompilationDto compilationDto);
 
     CompilationDto updateCompilation(long compId, UpdateCompilationRequest request);
 
     void deleteCompilation(long compId);
-
-    List<CompilationDto> getAllCompilations(Boolean pinned, int page, int size);
-
-    CompilationDto getCompilationById(long compId);
 }
