+++ core/event-server/src/main/java/ru/eventlink/compilation/service/CompilationPublicService.java
@@ -1,0 +1,11 @@
+package ru.eventlink.compilation.service;
+
+import ru.eventlink.dto.compilation.CompilationDto;
+
+import java.util.List;
+
+public interface CompilationPublicService {
+    List<CompilationDto> getAllCompilations(Boolean pinned, int page, int size);
+
+    CompilationDto getCompilationById(long compId);
+}
