+++ core/event-server/src/main/java/ru/eventlink/event/controller/EventPrivateController.java
@@ -9,7 +9,7 @@ import ru.eventlink.dto.event.*;
 import ru.eventlink.dto.requests.EventRequestStatusUpdateRequestDto;
 import ru.eventlink.dto.requests.EventRequestStatusUpdateResultDto;
 import ru.eventlink.dto.requests.ParticipationRequestDto;
-import ru.eventlink.event.service.EventService;
+import ru.eventlink.event.service.EventPrivateService;
 
 import java.util.List;
 
@@ -18,20 +18,20 @@ import java.util.List;
 @RequiredArgsConstructor
 @Slf4j
 public class EventPrivateController {
-    private final EventService eventService;
+    private final EventPrivateService eventPrivateService;
 
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
     public EventFullDto addEvent(@RequestBody @Valid NewEventDto newEventDto, @PathVariable long userId) {
         log.info("Received a POST request to add event {} from a user with an userId = {}", newEventDto, userId);
-        return eventService.addEvent(newEventDto, userId);
+        return eventPrivateService.addEvent(newEventDto, userId);
     }
 
     @GetMapping("/{eventId}")
     public EventFullDto findEventById(@PathVariable long userId,
                                       @PathVariable long eventId) {
         log.info("Received a GET request to find event by id {} from a user with an userId = {}", eventId, userId);
-        return eventService.findEventByUserIdAndEventId(userId, eventId);
+        return eventPrivateService.findEventByUserIdAndEventId(userId, eventId);
     }
 
     @GetMapping
@@ -39,7 +39,7 @@ public class EventPrivateController {
                                                 @RequestParam(defaultValue = "0") int page,
                                                 @RequestParam(defaultValue = "10") int size) {
         log.info("Received a GET request to find events by userId = {} page = {} size = {}", userId, page, size);
-        return eventService.findEventsByUser(userId, page, size);
+        return eventPrivateService.findEventsByUser(userId, page, size);
     }
 
     @PatchMapping("/{eventId}")
@@ -48,14 +48,14 @@ public class EventPrivateController {
                                     @PathVariable long eventId) {
         log.info("Received a PATCH request to update event with an eventId = {} from a user with an userId = {}, " +
                 "request body {}", eventId, userId, updateEventUserRequest);
-        return eventService.updateEvent(updateEventUserRequest, userId, eventId);
+        return eventPrivateService.updateEvent(updateEventUserRequest, userId, eventId);
     }
 
     @GetMapping("/{eventId}/requests")
     public List<ParticipationRequestDto> findRequestByEventId(@PathVariable long userId, @PathVariable long eventId) {
         log.info("Received a GET request to find request by event id = {} from a user with an userId = {}",
                 eventId, userId);
-        return eventService.findRequestByEventId(userId, eventId);
+        return eventPrivateService.findRequestByEventId(userId, eventId);
     }
 
     @PatchMapping("/{eventId}/requests")
@@ -66,12 +66,12 @@ public class EventPrivateController {
                                                                     @PathVariable long eventId) {
         log.info("Received a PATCH request to update request with an eventId = {} from a user with an userId = {}, " +
                 "request body {}", eventId, userId, updateRequests);
-        return eventService.updateRequestByEventId(updateRequests, userId, eventId);
+        return eventPrivateService.updateRequestByEventId(updateRequests, userId, eventId);
     }
 
     @GetMapping("/recommendations")
     public List<RecommendationsDto> findRecommendations(@PathVariable long userId) {
         log.info("Received a GET request to find recommendations");
-        return eventService.findRecommendations(userId);
+        return eventPrivateService.findRecommendations(userId);
     }
 }
